requirements.txt
cohere
flask
python-dotenv
requests 
pyngrok